requirements.txt
PySide6
pandas
openpyxl
numpy
scipy
matplotlib
reportlab
pyyaml
keyring
requests
pytest
xlsxwriter
pyinstaller
